Console: Move plugins/editor/utils/terminal.py to plugins/console/utils/ansihandler.py
