+++ spyder/plugins/console/utils/ansihandler.py
